pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sqlserver-doctor-mcp"
version = "0.1.0"
description = "MCP server for SQL Server tuning and diagnostics"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Jan Lindberg" }
]
dependencies = [
    "mcp>=1.2.1",
    "pyodbc>=5.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
sqlserver-doctor = "sqlserver_doctor.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
sqlserver_doctor = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
strict = true
